requirements.txt
#############################################
# RAG Repository Analyzer — Python runtime deps
#
# Lean, conflict‑free set of direct dependencies.
# Transitive dependencies are intentionally not pinned
# to avoid lock‑in and reduce resolver conflicts.
#
# If you need dev/test tooling, see requirements-dev.txt
#############################################

# Web API
fastapi==0.115.8
uvicorn[standard]==0.34.0  # includes uvloop, httptools, websockets, watchfiles
python-dotenv>=1.0.0,<2.0.0
pydantic==2.10.6

# Vector DB & embeddings
chromadb==0.6.3
chroma-hnswlib==0.7.6
mmh3==5.1.0
numpy>=1.26.2,<2.0.0  # cap <2 for langchain/chroma compatibility
tiktoken==0.8.0

# LangChain (splits in v0.3)
langchain==0.3.18
langchain-text-splitters==0.3.6

# Optional extras (uncomment as needed)
# openai>=1.0.0           # enable OpenAI embeddings/chat if OPENAI_API_KEY is set
# transformers>=4.44      # local HF embeddings (heavy)
# torch>=2.3              # required by transformers for local models
